Use comet to generate/enumerate the codeChunks.
Creates a notion of a "NotebookSession" that holds all the info for a
notebook, and a NotebookLines object that renders them to the page.

Notebooks are a collection of "codeChunk" objects; multi-line code fragments.

Add js and css specific to these codeChunks; double clicking a codeChunk
causes it to turn into an editable text area.